% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NA.R
\name{countna}
\alias{countna}
\title{This function goes through the whole data set, counts NA's and #'returns back a visualization as a bar plot if NA's are present. #'This is very useful #' when working with an untidy data frame to #'first see how many NAs and which variables have the missing data}
\usage{
countna(data)
}
\arguments{
\item{data}{tibble or data frame that you want use to test for missing NA's}
}
\value{
a bar graph of {data} will plotted NA's or a message indicating no NA's
}
\description{
This function goes through the whole data set, counts NA's and #'returns back a visualization as a bar plot if NA's are present. #'This is very useful #' when working with an untidy data frame to #'first see how many NAs and which variables have the missing data
}
\examples{
\dontrun{countna(datateachr::vancouver_trees)}
\dontrun{.countna(datateachr::apt_buildings)}
}
